refactoring func CONVERT  Everything is fine!!!!!!!
