+++ MovieQuiz/Presentation/MovieQuizPresenter.swift
@@ -1,0 +1,35 @@
+//
+//  MovieQuizPresenter.swift
+//  MovieQuiz
+//
+//  Created by Jedi on 22.06.2023.
+//
+import UIKit
+import Foundation
+
+final class MovieQuizPresenter {
+    
+    let questionsAmount: Int = 10
+        private var currentQuestionIndex: Int = 0
+        
+        func isLastQuestion() -> Bool {
+            currentQuestionIndex == questionsAmount - 1
+        }
+        
+        func resetQuestionIndex() {
+            currentQuestionIndex = 0
+        }
+        
+        func switchToNextQuestion() {
+            currentQuestionIndex += 1
+        }
+    
+     func convert(model: QuizQuestion) -> QuizStepViewModel {
+        let questionStep = QuizStepViewModel(
+            image: UIImage(data: model.image) ?? UIImage(),
+            question: model.text,
+            questionNumber: "\(currentQuestionIndex + 1)/\(questionsAmount)")
+        return questionStep
+    }
+    
+}
